requirements.txt
# SAR Water Detection Lab - Complete Requirements
# ==================================================

# Core Framework
streamlit>=1.31.0
streamlit-drawable-canvas>=0.9.3

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Geospatial
rasterio>=1.3.9
GDAL>=3.6.0
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0

# Image Processing & Visualization
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Machine Learning
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0

# Deep Learning Losses (Optional)
monai>=1.3.0

# Wavelets
PyWavelets>=1.4.1

# Utilities
tqdm>=4.66.0
click>=8.1.0
python-dotenv>=1.0.0

# Development & Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation (Optional)
mkdocs>=1.5.0
mkdocs-material>=9.5.0
